+++ src/assault/mod.rs
@@ -4,5 +4,6 @@ pub mod evaluator;
 pub mod factory;
 pub mod service;
 
+pub mod metrics;
 pub mod reportable;
 pub mod worker;
